fix hashing of edge_descriptor for OpenMesh
